requirements.txt
tensorflow==2.9.1
opencv-contrib-python
pyrealsense2
depthai
open3d
matplotlib
scipy
pyyaml
tensorflow_addons
segmentation_models
focal_loss
begins
silence_tensorflow
